--- pkg/models/common_category.go
@@ -1,8 +1,0 @@
-package models
-
-// CommonCategory represents a deduplicated top-level template category
-// and its deduplicated, alphabetized subcategory names.
-type CommonCategory struct {
-	Name          string   `json:"name"`
-	SubCategories []string `json:"subCategories"`
-}
