+++ pkg/sqlqueries/SQL/mto/common_milestone/get_common_categories.sql
