+++ query_examples/MINT/Models to Operations Matrix/Common Milestones/Get Common Categories.bru
